Split endpoints module into smaller submodules
